+++ com.unity.shadergraph/Editor/Serialization/JsonObject.cs
@@ -14,6 +14,12 @@ namespace UnityEditor.ShaderGraph.Serialization
         [SerializeField]
         string m_ObjectId = Guid.NewGuid().ToString("N");
 
+        public string ObjectIdUnsafe
+        {
+            get { return m_ObjectId; }
+            set { m_ObjectId = value; } // TODO: Maybe throw exception if the assigned Id is not a valid guid.
+        }
+
         public string objectId => m_ObjectId;
 
         public bool objectIdIsEmpty => m_ObjectId.Equals(emptyObjectId);